requirements.txt
python-binance==1.0.19
pandas==2.1.1
numpy==1.24.3
ta==0.10.2
tensorflow==2.13.0
scikit-learn==1.3.0
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
sqlalchemy==2.0.21
pandas-ta==0.3.14b0
ccxt==4.4.45
python-telegram-bot==20.5
loguru==0.7.2